content/canvas/test/test_2d.state.saverestore.globalCompositeOperation.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.state.saverestore.globalCompositeOperation</title>
<!-- Testing: save()/restore() works for globalCompositeOperation -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

// Test that restore() undoes any modifications
var old = ctx.globalCompositeOperation;
ctx.save();
ctx.globalCompositeOperation = "copy";
ctx.restore();
ok(ctx.globalCompositeOperation === old, "ctx.globalCompositeOperation === old");

// Also test that save() doesn't modify the values
ctx.globalCompositeOperation = "copy";
old = ctx.globalCompositeOperation;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against d instead of against "copy"
ctx.save();
ok(ctx.globalCompositeOperation === old, "ctx.globalCompositeOperation === old");
ctx.restore();

SimpleTest.finish();

});
</script>
